Feat: added initial-info page
